ibc: light clients restructure
