+++ docs/ibc/overview.md
@@ -35,7 +35,7 @@ chains. The supported IBC clients are:
 
 * [Solo Machine light client](https://github.com/cosmos/cosmos-sdk/tree/master/x/ibc/light-clients/solomachine): devices such as phones, browsers, or laptops.
 * [Tendermint light client](https://github.com/cosmos/cosmos-sdk/tree/master/x/ibc/07-tendermint): The default for SDK-based chains,
-* [Localhost (loopback) client](https://github.com/cosmos/cosmos-sdk/tree/master/x/ibc/09-localhost): Useful for
+* [Localhost (loopback) client](https://github.com/cosmos/cosmos-sdk/tree/master/x/ibc/light-clients/09-localhost): Useful for
 testing, simulation and relaying packets to modules on the same application.
 
 ### [Connections](https://github.com/cosmos/cosmos-sdk/tree/master/x/ibc/03-connection)
