+++ proto/ibc/localhost/localhost.proto
@@ -4,7 +4,7 @@ package ibc.localhost;
 import "gogoproto/gogo.proto";
 import "ibc/client/client.proto";
 
-option go_package = "github.com/cosmos/cosmos-sdk/x/ibc/09-localhost/types";
+option go_package = "github.com/cosmos/cosmos-sdk/x/ibc/light-clients/09-localhost/types";
 
 // ClientState defines a loopback (localhost) client. It requires (read-only)
 // access to keys outside the client prefix.
