+++ x/ibc/light-clients/09-localhost/module.go
@@ -1,7 +1,7 @@
 package localhost
 
 import (
-	"github.com/cosmos/cosmos-sdk/x/ibc/09-localhost/types"
+	"github.com/cosmos/cosmos-sdk/x/ibc/light-clients/09-localhost/types"
 )
 
 // Name returns the IBC client name
